requirements.txt
faiss-cpu
groq
langchain-groq
langchain-google-genai
langchain
streamlit
langchain-community
python-dotenv
pymupdf
fpdf